ANCHOR SCROLL: WIP - feat($anchorScroll): add support for configurable scroll offset
Updated the docs and implemented the dynamic scroll offset in the docs app
in a more proper way (introduced the `scrollOffset` module as well).
There seems to be an issue with the current implementation, as
`window.scrollBy()` does not work when the "scrolled" element is lower in
the DOM hierarchy.